% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_brands.R
\name{clean_brands}
\alias{clean_brands}
\title{This function accepts a data frame argument
It searches for any column header having to do with brands
and then will replace the strings "PRIVATE LABEL" and "NATIONAL" 
with "PRIVATE BRAND" and "BRANDED", respectively.
Note: This function is insensitive to case in header names, but 
IS CASE SENSITIVE to brand names.}
\usage{
clean_brands(x)
}
\description{
This function accepts a data frame argument
It searches for any column header having to do with brands
and then will replace the strings "PRIVATE LABEL" and "NATIONAL" 
with "PRIVATE BRAND" and "BRANDED", respectively.
Note: This function is insensitive to case in header names, but 
IS CASE SENSITIVE to brand names.
}
