% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_headers.R
\name{clean_headers}
\alias{clean_headers}
\title{This function accepts a data frame argument
and cleans up the headers
Only works with data exported as xls
clean_headers}
\usage{
clean_headers(x)
}
\description{
This function accepts a data frame argument
and cleans up the headers
Only works with data exported as xls
clean_headers
}
