% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_periods.R
\name{clean_periods}
\alias{clean_periods}
\title{This function parses the period column returned by AOD
and returns two columns, one indicating the type of period
and the other the week ending for the period, formatted as a date
and cleans up the headers.}
\usage{
clean_periods(data, period_col)
}
\description{
This function parses the period column returned by AOD
and returns two columns, one indicating the type of period
and the other the week ending for the period, formatted as a date
and cleans up the headers.
}
\examples{
data <- clean_periods(data, "Periods")
}
